Moved to zsh and small improvements.
- Moved from Bash to Zsh as the file globbing is better
- Added a check to skip the move if an embedded object was not found, to stop premature exits

+++ extract.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/local/bin/zsh
 
 ## A simple tool to strip the first attachement from an Excel file.
 
@@ -12,12 +12,12 @@ TARGET=$2
 COUNT=0
 
 ## Bail out if the source or destination are bad
-if [[ ! -d $FILES ]]; then
+if [[ ! -d "$FILES" ]]; then
 	echo "Source directory $FILES does not exist"
 	exit 1
 fi
 
-if [[ ! -d $TARGET ]]; then
+if [[ ! -d "$TARGET" ]]; then
         echo "Target directory $TARGET does not exist"
         exit 1
 fi
@@ -25,7 +25,7 @@ fi
 ##Create a temp directory to work in
 WORKDIR=$(mktemp -d)
 
-if [[ ! $WORKDIR || ! -d $WORKDIR ]]; then
+if [[ ! "$WORKDIR" || ! -d "$WORKDIR" ]]; then
 	echo "could not create $WORKDIR"
 	exit 1
 fi
@@ -39,10 +39,14 @@ function cleanup {
 trap cleanup EXIT
 
 ## Loop through all the Excel files in the directory, and extract the first attachment.
-for f in $FILES/*.xlsx; do
-	unzip -q -d "$WORKDIR/$COUNT" $f || { echo "could not extract $f"; exit 1; }
-	echo "moving $WORKDIR/$COUNT/xl/embeddings/oleObject1.bin to $TARGET/invoice_$COUNT.pdf"
-	mv "$WORKDIR/$COUNT/xl/embeddings/oleObject1.bin" "$TARGET/invoice_$COUNT.pdf" || { echo "could not move file to $TARGET"; exit 1; }
+for f in  $FILES/**/*.xlsx ; do
+	unzip -q -d "$WORKDIR/$COUNT" "$f" 
+	if [[ -f "$WORKDIR/$COUNT/xl/embeddings/oleObject1.bin" ]]; then
+		echo "moving $WORKDIR/$COUNT/xl/embeddings/oleObject1.bin to $TARGET/invoice_$COUNT.pdf"
+		mv "$WORKDIR/$COUNT/xl/embeddings/oleObject1.bin" "$TARGET/invoice_$COUNT.pdf" || { echo "could not move file to $TARGET"; exit 1; }
+	else
+		echo "no PDF found in $f. Skipping"
+	fi	
 	((COUNT++))
 done
